requirements.txt
beautifulsoup4==4.11.1
moviepy==2.0.0.dev2
openai_whisper==20230314
PyAutoGUI==0.9.53
requests==2.28.2
selenium==4.9.0
undetected_chromedriver==3.4.6
